Refactor code to avoid cyclic reference
