+++ src/server.ts
@@ -1,10 +1,10 @@
 import './utils/env';
 import {PORT} from './utils/env';
-import boltApp from './bolt/app';
 import expressApp from './express/app';
+import slackBotApp from './bolt/app';
 
 (async () => {
-    await boltApp.start();
+    await slackBotApp.start();
     return expressApp.listen(PORT);
 })().then(() => {
     console.log('⚡️ Bolt boltApp is running!');
